-All grabs require less mashing to escape, but more mashing has decreased gain
Fixed cherrypicking issues

+++ Content/Projectiles/Masomode/DeerclopsHand.cs
@@ -32,7 +32,7 @@ namespace FargowiltasSouls.Content.Projectiles.Masomode
         }
 
         bool spawned;
-        int oldMash;
+        float oldMash;
 
         public override void AI()
         {
